port theme, remove old main and update deps
